+++ ha-sip/src/command_handler.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import collections.abc
 import sys
 import os
-from typing import Optional
+from typing import Optional, List
 
 import pjsua2 as pj
 
@@ -13,10 +13,10 @@ import command_client
 import ha
 import state
 import utils
-from call_state_change import CallStateChange
 from constants import DEFAULT_RING_TIMEOUT
 from event_sender import EventSender
 from log import log
+from post_action import PostActionHangup
 
 
 class CommandHandler(object):
@@ -43,12 +43,15 @@ class CommandHandler(object):
     def is_active(self, caller_id: str) -> bool:
         return self.call_state.is_active(caller_id)
 
-    def on_state_change(self, state_change: CallStateChange, caller_id: str, new_call: call.Call) -> None:
-        self.call_state.on_state_change(state_change, caller_id, new_call)
+    def register_call(self, callback_id: str, new_call: call.Call, additional_ids: List[str]) -> None:
+        self.call_state.register_call(callback_id, new_call, additional_ids)
+
+    def forget_call(self, callback_id: str) -> None:
+        self.call_state.forget_call(callback_id)
 
     def handle_command(self, command: command_client.Command, from_call: Optional[call.Call]) -> None:
         if not isinstance(command, collections.abc.Mapping):
-            log(None, 'Error: Not an object: %s' % command)
+            log(None, f'Error: Not an object: {command}')
             return
         verb = command.get('command')
         number_unknown_type = command.get('number')
@@ -62,18 +65,18 @@ class CommandHandler(object):
                 if (not domain) or (not service):
                     log(None, 'Error: one of domain or service was not provided')
                     return
-                log(None, 'Calling home assistant service on domain %s service %s with entity %s' % (domain, service, entity_id))
+                log(None, f'Calling home assistant service on domain {domain} service {service} with entity {entity_id}')
                 try:
                     ha.call_service(self.ha_config, domain, service, entity_id, service_data)
                 except Exception as e:
-                    log(None, 'Error calling home-assistant service: %s' % e)
+                    log(None, f'Error calling home-assistant service: {e}')
             case 'dial':
                 if not number:
                     log(None, 'Error: Missing number for command "dial"')
                     return
-                log(None, 'Got "dial" command for %s' % number)
+                log(None, f'Got "dial" command for {number}')
                 if self.is_active(number):
-                    log(None, 'Warning: call already in progress: %s' % number)
+                    log(None, f'Warning: call already in progress: {number}')
                     return
                 menu = command.get('menu')
                 ring_timeout = utils.convert_to_float(command.get('ring_timeout'), DEFAULT_RING_TIMEOUT)
@@ -85,7 +88,7 @@ class CommandHandler(object):
                 if not number:
                     log(None, 'Error: Missing number for command "hangup"')
                     return
-                log(None, 'Got "hangup" command for %s' % number)
+                log(None, f'Got "hangup" command for {number}')
                 if not self.is_active(number):
                     self.call_not_in_progress_error(number)
                     return
@@ -95,7 +98,7 @@ class CommandHandler(object):
                 if not number:
                     log(None, 'Error: Missing number for command "answer"')
                     return
-                log(None, 'Got "answer" command for %s' % number)
+                log(None, f'Got "answer" command for {number}')
                 if not self.is_active(number):
                     self.call_not_in_progress_error(number)
                     return
@@ -139,13 +142,13 @@ class CommandHandler(object):
                     return
                 digits = command.get('digits')
                 method = command.get('method', 'in_band')
-                if (method != 'in_band') and (method != 'rfc2833') and (method != 'sip_info'):
+                if method not in ('in_band', 'rfc2833', 'sip_info'):
                     log(None, 'Error: method must be one of in_band, rfc2833, sip_info')
                     return
                 if not digits:
                     log(None, 'Error: Missing digits for command "send_dtmf"')
                     return
-                log(None, 'Got "send_dtmf" command for %s' % number)
+                log(None, f'Got "send_dtmf" command for {number}')
                 if not self.is_active(number):
                     self.call_not_in_progress_error(number)
                     return
@@ -165,6 +168,13 @@ class CommandHandler(object):
                     return
                 cache_audio = command.get('cache_audio') or False
                 wait_for_audio_to_finish = command.get('wait_for_audio_to_finish') or False
+                match command.get('post_action'):
+                    case 'hangup':
+                        current_call.scheduled_post_action = PostActionHangup(action="hangup")
+                    case 'noop':
+                        pass
+                    case _:
+                        log(None, 'Only post_action "hangup" is supported. Assuming noop.')
                 current_call.play_audio_file(audio_file, cache_audio, wait_for_audio_to_finish)
             case 'play_message':
                 if not number:
@@ -178,9 +188,19 @@ class CommandHandler(object):
                 if not message:
                     log(None, 'Error: Missing parameter "message" for command "play_message"')
                     return
+                handle_as_template = command.get('handle_as_template')
+                if handle_as_template:
+                    message = ha.render_template(current_call.ha_config, message)
                 tts_language = command.get('tts_language') or self.ha_config.tts_config['language']
                 cache_audio = command.get('cache_audio') or False
                 wait_for_audio_to_finish = command.get('wait_for_audio_to_finish') or False
+                match command.get('post_action'):
+                    case 'hangup':
+                        current_call.scheduled_post_action = PostActionHangup(action="hangup")
+                    case 'noop':
+                        pass
+                    case _:
+                        log(None, 'Only post_action "hangup" is supported. Assuming noop.')
                 current_call.play_message(message, tts_language, cache_audio, wait_for_audio_to_finish)
             case 'stop_playback':
                 if not number:
@@ -220,8 +240,8 @@ class CommandHandler(object):
                 self.end_point.libDestroy()
                 sys.exit(0)
             case _:
-                log(None, 'Error: Unknown command: %s' % verb)
+                log(None, f'Error: Unknown command: {verb}')
 
     def call_not_in_progress_error(self, number: str):
-        log(None, 'Warning: call not in progress: %s' % number)
+        log(None, f'Warning: call not in progress: {number}')
         self.call_state.output()