+++ ha-sip/src/main.py
@@ -1,27 +1,29 @@
 #!/usr/bin/env python3
 import asyncio
-import os
 import faulthandler
-from typing import Optional
+import os
 import sys
+from typing import Optional, Any
 
 import yaml
+
 import account
 import call
+import config
 import ha
 import incoming_call
+import mqtt
 import options_global
 import options_sip
 import sip
 import state
 import utils
-import mqtt
 from command_client import CommandClient
 from command_handler import CommandHandler
 from event_sender import EventSender
 from ha import TtsConfigFromEnv
 from log import log
-import config
+
 
 def handle_command_list(command_client: CommandClient, command_handler: CommandHandler) -> None:
     command_list = command_client.get_command_list()
@@ -36,10 +38,10 @@ def load_menu_from_file(file_name: Optional[str], sip_account_index: int) -> Opt
     try:
         with open(file_name) as stream:
             content = yaml.safe_load(stream)
-            log(sip_account_index, 'Loaded menu for incoming call from "%s".' % file_name)
+            log(sip_account_index, f'Loaded menu for incoming call from "{file_name}".')
             return content
     except BaseException as e:
-        log(sip_account_index, 'Error loading menu for incoming call: %s' % e)
+        log(sip_account_index, f'Error loading menu for incoming call: {e}')
         return None
 
 
@@ -47,7 +49,7 @@ def get_name_server(raw_name_server: str):
     name_server = [ns.strip() for ns in raw_name_server.split(",")]
     name_server_without_empty = [ns for ns in name_server if ns]
     if name_server_without_empty:
-        log(None, 'Setting name server: %s' % name_server)
+        log(None, f'Setting name server: {name_server}')
     return name_server_without_empty
 
 
@@ -58,7 +60,7 @@ def get_cache_dir(raw_cache_dir: str) -> Optional[str]:
     if not os.path.isdir(raw_cache_dir):
         log(None, 'Error: Cache directory not found.')
         return None
-    log(None, "Found cache directory '%s'" % raw_cache_dir)
+    log(None, f"Found cache directory '{raw_cache_dir}'")
     return raw_cache_dir
 
 
@@ -139,9 +141,9 @@ def main():
             is_first_enabled_account = False
     mqtt_mode = config.COMMAND_SOURCE.lower().strip() == 'mqtt'
     mqtt_client = mqtt.create_client_and_connect(command_handler) if mqtt_mode else None
-    def trigger_webhook(event: ha.WebhookEvent, webhook_id: Optional[str] = None):
+    def trigger_webhook(event: Any, webhook_id: Optional[str] = None):
         ha.trigger_webhook(ha_config, event, webhook_id)
-    def send_mqtt_event(event: ha.WebhookEvent, webhook_id: Optional[str] = None):
+    def send_mqtt_event(event: Any, webhook_id: Optional[str] = None):
         if mqtt_client:
             mqtt_client.send_event(event)
     event_sender.register_sender(trigger_webhook)